% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_source.R
\name{available_providers}
\alias{available_providers}
\title{Name providers available for a given version}
\usage{
available_providers(version = latest_version(), db = td_connect())
}
\arguments{
\item{version}{snapshot version, defaults to the latest available}

\item{db}{a connection from \code{\link[=td_connect]{td_connect()}}}
}
\value{
a data.frame of \code{provider} and the \code{schema}s published for it
}
\description{
Name providers available for a given version
}
\examples{
\donttest{
available_providers()
}
}
