% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_common.R
\name{by_common}
\alias{by_common}
\title{Look up taxonomic information by common name}
\usage{
by_common(name, provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb",
  "slb", "wd", "ott", "iucn"), collect = TRUE, ignore_case = TRUE,
  db = td_connect())
}
\arguments{
\item{name}{a character vector of common (vernacular English) names, e.g. "Humans"}

\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{ignore_case}{should we ignore case (capitalization) in matching names?
default is \code{TRUE}.}

\item{db}{a connection to the taxadb database. See details.}
}
\value{
a data.frame in the Darwin Core tabular format containing the
matching taxonomic entities.
}
\description{
Look up taxonomic information by common name
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }

by_common("Angolan Giraffe")

}

}
\seealso{
Other filter_by: \code{\link{by_id}},
  \code{\link{by_name}}, \code{\link{by_rank}},
  \code{\link{filter_by}}
}
\concept{filter_by}
