% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_id.R
\name{by_id}
\alias{by_id}
\title{Get all members (descendants) of a given rank level}
\usage{
by_id(id, provider = known_providers, type = c("taxonID",
  "acceptedNameUsageID"), collect = TRUE, db = td_connect())
}
\arguments{
\item{id}{taxonomic id, in prefix format}

\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{type}{id type.  Can be taxonID or \code{acceptedNameUsageID}. Use \code{acceptedNameUsageID}
to return all rows for which this ID is the accepted ID, including both synonyms and
and accepted names (since both all synonyms of a name share the same \code{acceptedNameUsageID}.)
Use taxonID (default) to only return those rows for which the Scientific name
corresponds to the taxonID.

Some providers (e.g. ITIS) assign taxonIDs to synonyms, most others only assign IDs to
accepted names.  In the latter case, this means requesting \code{taxonID} will only match
accepted names, while requesting matches to the \code{acceptedNameUsageID} will also return
any known synonyms.  See examples.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxadb database. See details.}
}
\value{
a data.frame with id and name of all matching species
}
\description{
Get all members (descendants) of a given rank level
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }

by_id(c("ITIS:1077358", "ITIS:175089"))
by_id("ITIS:1077358", type="acceptedNameUsageID")

}
}
