% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_rank.R
\name{by_rank}
\alias{by_rank}
\title{Get all members (descendants) of a given rank level}
\usage{
by_rank(name, rank, provider = c("itis", "ncbi", "col", "tpl", "gbif",
  "fb", "slb", "wd", "ott", "iucn"), collect = TRUE,
  ignore_case = TRUE, db = td_connect())
}
\arguments{
\item{name}{taxonomic scientific name (e.g. "Aves")}

\item{rank}{taxonomic rank name. (e.g. "class")}

\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{ignore_case}{should we ignore case (capitalization) in matching names?
default is \code{TRUE}.}

\item{db}{a connection to the taxadb database. See details.}
}
\value{
a data.frame in the Darwin Core tabular format containing the
matching taxonomic entities.
}
\description{
Get all members (descendants) of a given rank level
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }

by_rank("Aves", "class")

}

}
\seealso{
Other filter_by: \code{\link{by_common}},
  \code{\link{by_id}}, \code{\link{by_name}},
  \code{\link{filter_by}}, \code{\link{name_contains}}
}
\concept{filter_by}
