% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Return the full classification hierarchy for requested species or ids}
\usage{
classification(species = NULL, id = NULL, provider = known_providers,
  collect = TRUE, db = td_connect())
}
\arguments{
\item{species}{a character vector of species names, typically
specified as (\code{Genus species} or \code{Genus species epithet})}

\item{id}{alternately users can provide a vector of species ids.
IDs must be prefixed matching the requested provider.  See \code{id}
column returned by most \code{taxadb} functions for examples.}

\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxadb database. See details.}
}
\value{
a data.frame with one row for each requested species,
giving the species id, and column for each of the unique-rank
levels of the species.  Note that the different authorities recognize
a range of different ranks.
}
\description{
Return the full classification hierarchy for requested species or ids
}
\details{
Some authorities recognize multiple values of the same rank
level, (i.e. a species may be assigned to two different "suborders").
In this case, only one has been included in the "hierarchy" schema.
Likewise, some authorities (i.e. NCBI) use some unnamed rank levels
(i.e. a scientific name is associated with a clade, but the clade is
not associated with any traditionally recognized rank). In these cases,
see the "long" schema for more complete classification.
}
