% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_db}
\alias{create_db}
\alias{create_taxadb}
\title{create a local taxonomic database}
\usage{
create_db(authorities = "itis", schema = "hierarchy",
  overwrite = FALSE, lines = 1e+06, path = Sys.getenv("TAXALD_HOME",
  file.path(path.expand("~"), ".taxald")))
}
\arguments{
\item{authorities}{a list (character vector) of authorities to be included in the
database. By default, will install \code{itis}.  See details for a list of recognized
authorities. Use \code{authorities="all"} to install all available authorities automatically.}

\item{schema}{format of the database to import.}

\item{overwrite}{Should we overwrite existing tables? Default is \code{FALSE}.  Change
to "ask" for interactive interface, or \code{TRUE} to force overwrite (i.e. updating
a local database upon new release.)}

\item{lines}{number of lines that can be safely read in to memory at once. Leave
at default or increase for faster importing if you have plenty of spare RAM.}

\item{path}{a location on your computer where the database should be installed.
By default, will install to \code{.taxald} in your home directory.}
}
\value{
path where database has been installed (invisibly)
}
\description{
create a local taxonomic database
}
\details{
Authorities recognized by taxald are:
\itemize{
\item \code{itis}
\item \code{ncbi}
\item \code{col}
\item \code{tpl}
\item \code{gbif}
\item \code{fb} FishBase, \url{http://fishbase.org}
\item \code{slb}, SeaLifeBase, \url{http://sealifebase.org}
\item \code{wd}, wikidata
}
}
\examples{
\dontrun{
  # tmp <- tempdir()
  # create_db(tmp, authorities = "itis")

}
}
