% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize.R
\name{get_uid}
\alias{get_uid}
\title{Get NCBI Identifiers
A drop-in replacement for \code{[taxize::get_uid()]}}
\usage{
get_uid(sciname, ...)
}
\arguments{
\item{a}{scientific name (or vector of scientific names)}
}
\description{
Get NCBI Identifiers
A drop-in replacement for \code{[taxize::get_uid()]}
}
