% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_source.R
\name{latest_version}
\alias{latest_version}
\title{The most recent taxadb snapshot version}
\usage{
latest_version(db = td_connect())
}
\arguments{
\item{db}{a connection from \code{\link[=td_connect]{td_connect()}}}
}
\value{
the latest available version, as a character string
}
\description{
The most recent taxadb snapshot version
}
\examples{
\donttest{
latest_version()
}
}
