% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_source.R
\name{taxadb_dir}
\alias{taxadb_dir}
\title{Show the local taxadb directory}
\usage{
taxadb_dir()
}
\value{
path to the local taxadb data directory
}
\description{
Show the local taxadb directory
}
\details{
Local snapshots downloaded by \code{\link[=td_download]{td_download()}} are stored here.
Override with the \code{TAXADB_HOME} environment variable.
}
\examples{
taxadb_dir()
}
