% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_source.R
\name{taxadb_uri}
\alias{taxadb_uri}
\title{Locate the Parquet files backing a taxadb table}
\usage{
taxadb_uri(
  provider = getOption("taxadb_default_provider", "itis"),
  schema = c("dwc", "common"),
  version = latest_version(),
  local = NULL
)
}
\arguments{
\item{provider}{from which provider should the hierarchy be returned?
Default is 'itis', which can also be configured using \verb{options(default_taxadb_provider=...")}.
See \verb{[td_create]} for a list of recognized providers.}

\item{schema}{One of "dwc" (for Darwin Core data) or "common"
(for the Common names table.)}

\item{version}{Which version of the taxadb provider database should we use?
defaults to latest.  See \code{\link[=available_versions]{available_versions()}} for details.}

\item{local}{should we return the path to a local snapshot?  By default
a local copy is used when one is present (see \code{\link[=td_download]{td_download()}}), and the
remote snapshot is streamed otherwise.}
}
\value{
a glob pattern (or file path) that \code{duckdb} can read
}
\description{
Locate the Parquet files backing a taxadb table
}
\examples{
taxadb_uri("itis_test")
}
