% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_home.R
\name{td_home}
\alias{td_home}
\title{td_home}
\usage{
td_home(dbdir = Sys.getenv("TAXALD_HOME"), create = TRUE)
}
\arguments{
\item{dbdir}{a location on your computer where the database should be installed.
By default, will look for a location given by \code{Sys.getenv("TAXALD_HOME")} if not specified.}

\item{create}{should the location be created if no directory exits?}
}
\value{
dbdir path that will be used.
}
\description{
td_home
}
