% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_source.R
\name{available_versions}
\alias{available_versions}
\title{Versions of the taxadb data available}
\usage{
available_versions(db = td_connect())
}
\arguments{
\item{db}{a connection from \code{\link[=td_connect]{td_connect()}}}
}
\value{
a character vector of available snapshot versions
}
\description{
Versions of the taxadb data available
}
\examples{
\donttest{
available_versions()
}
}
