% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_create.R
\name{available_versions}
\alias{available_versions}
\title{List available releases}
\usage{
available_versions()
}
\description{
taxadb uses pre-computed cache files that are released on an annual
version schedule.
}
\examples{
available_versions()
}
