% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create_taxadb}
\alias{create_taxadb}
\title{create a local taxonomic database}
\usage{
create_taxadb(path = fs::path(fs::path_home(), ".taxald"),
  authorities = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb", "wiki"))
}
\arguments{
\item{path}{a location on your computer where the database should be installed.
By default, will install to \code{.taxald} in your home directory.}

\item{authorities}{a list (character vector) of authorities to be included in the
database. By default, will install all authorities.  Choose a subset for a faster
install.}
}
\value{
path where database has been installed (invisibly)
}
\description{
create a local taxonomic database
}
\examples{
\dontrun{
  tmp <- tempdir()
  create_taxadb(tmp, authorities = "itis")

}
}
