% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{get_ids}
\usage{
get_ids(names, db = known_providers, format = c("prefix", "bare",
  "uri"), taxadb_db = td_connect(), ...)
}
\arguments{
\item{names}{a list of scientific names (which may
include higher-order ranks in most authorities).}

\item{db}{abbreviation code for the provider.  See details.}

\item{format}{Format for the returned identifier, one of
\itemize{
\item \code{prefix} (e.g. \code{NCBI:9606}, the default), or
\item \code{bare} (e.g. \code{9606}, used in \code{taxize::get_ids()}),
\item \code{uri} (e.g.
\code{http://ncbi.nlm.nih.gov/taxonomy/9606}).
}}

\item{taxadb_db}{Connection to from \code{[td_connect]()}.}

\item{...}{additional arguments passed to \code{ids()}}
}
\description{
A drop-in replacement for \code{[taxize::get_ids()]}
}
\details{
Note that some taxize authorities: \code{nbn}, \code{tropicos}, and \code{eol},
are not recognized by taxadb and will throw an error here. Meanwhile,
taxadb recognizes several authorities not known to \code{[taxize::get_ids()]}.
Both include \code{itis}, \code{ncbi}, \code{col}, and \code{gbif}.

Like all taxadb functions, this function will run
fastest if a local copy of the provider is installed in advance
using \code{[td_create()]}.
}
\examples{
\donttest{
get_ids("Homo sapiens")
get_ids(c("Homo sapiens", "Mammalia"), format = "prefix")
get_ids("Homo sapiens", db= "ncbi", format = "uri")
}
}
