% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxize.R
\name{get_ids}
\alias{get_ids}
\title{get_ids}
\usage{
get_ids(names, db = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd"), format = c("bare", "prefix", "uri"), taxald_db = td_connect(),
  ...)
}
\arguments{
\item{names}{a list of scientific names (which may
include higher-order ranks in most authorities).}

\item{db}{abbreviation code for the authority.  See details.}

\item{format}{Format for the returned: bare identifier, one of
\itemize{
\item \code{bare} (e.g. \code{9606}, default, matching \code{taxize::get_ids()}),
\item \code{prefix} (e.g. \code{NCBI:9606}), or
\item \code{uri} (e.g.
\code{https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=9606}).
}}

\item{taxald_db}{Connection to from \code{[td_connect]()}.}

\item{...}{additional arguments passed to \code{ids()}}
}
\description{
A drop-in replacement for \code{[taxize::get_ids()]}
}
\details{
Note that some taxize authorities: \code{nbn}, \code{tropicos}, and \code{eol},
are not recognized by taxald and will throw an error here. Meanwhile,
taxald recognizes several authorities not known to \code{[taxize::get_ids()]}.
Both include \code{itis}, \code{ncbi}, \code{col}, and \code{gbif}.

Like all taxald functions, this function will run
fastest if a local copy of the authority is installed in advance
using \code{[td_create()]}.
}
\examples{
\donttest{
get_ids("Homo sapiens")
get_ids(c("Homo sapiens", "Mammalia"), format = "prefix")
get_ids("Homo sapiens", db= "ncbi", format = "uri")
}
}
