% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{get_names}
\usage{
get_names(ids, db = known_providers, format = c("guess", "prefix",
  "bare", "uri"), taxadb_db = td_connect(), ...)
}
\arguments{
\item{ids}{a list of taxonomic identifiers.}

\item{db}{abbreviation code for the provider.  See details.}

\item{format}{Format for the input identifier, one of
\itemize{
\item \code{guess} Will determine the prefix automatically, but will be
slower than if you specify this in advance!
\item \code{prefix} (e.g. \code{NCBI:9606}), Preferred (fastest) format.
\item \code{bare} (e.g. \code{9606}), (But must mach provider \code{db}!)
\item \code{uri} (e.g. \code{http://ncbi.nlm.nih.gov/taxonomy/9606}).
}}

\item{taxadb_db}{Connection to from \code{[td_connect]()}.}

\item{...}{additional arguments passed to \code{ids()}}
}
\description{
Translate identifiers into scientific names
}
\details{
Like all taxadb functions, this function will run
fastest if a local copy of the provider is installed in advance
using \code{[td_create()]}.
}
\examples{
\donttest{
get_names("180092")
get_names(c("ITIS:180092" "ITIS:179913"))
get_names(c("ITIS:180092", "ITIS:179913"), format = "prefix")
}
}
