% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Return taxonomic identifiers from a given namespace}
\usage{
ids(name = NULL, authority = c("itis", "ncbi", "col", "tpl", "gbif",
  "fb", "slb", "wd"), collect = TRUE, db = td_connect())
}
\arguments{
\item{name}{a character vector of species names, e.g. "Homo sapiens"
(Most but not all authorities can also return ids for higher-level
taxonomic names).}

\item{authority}{from which authority should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxadb database. See details.}
}
\value{
a data.frame with columns of \code{id}, scientific
\code{name}, and \code{rank}, and \code{accepted_id} (if data includes synonyms)
and a row for each species name queried.
}
\description{
Return taxonomic identifiers from a given namespace
}
\details{
Some authorities (ITIS, COL, FB/SLB, NCBI) provide synonyms as well.  These
are included in the name list search.  The data.frame returned will then
include an \code{accepted_id} column, providing the synonym for the requested table.
In this case, the \code{id} column corresponds to the id for either the name column --
that is, if the name is a synonym, this is the id for the synonym; if the name
is accepted, this id is the same as the accepted id.  NCBI does not issue ids
for known synonyms, so the id is missing for synonym names in this case.
}
