% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Return taxonomic identifiers from a given namespace}
\usage{
ids(name = NULL, authority = c("itis", "ncbi", "col", "tpl", "gbif",
  "fb", "slb", "wd"), collect = TRUE, db = td_connect())
}
\arguments{
\item{name}{a character vector of species names, e.g. "Homo sapiens"
(Most but not all authorities can also return ids for higher-level
taxonomic names).}

\item{authority}{from which authority should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxald database. See details.}

\item{pull}{logical, should we pull out the id column or return the full table?}
}
\value{
a data.frame with columns of \code{id}, scientific
\code{name}, and \code{rank} and a row for each species name queried.
}
\description{
Return taxonomic identifiers from a given namespace
}
\details{
NOTE on matching synonyms: Some authorities (like ITIS) issue separate synonym_ids
corresponding to synonym names. The \code{ids()} function \emph{does not} return synonym ids.
Rather, if a name is recognized as a synonym, we will look up the appropriate
accepted name and the ID associated with the accepted name and return that.
}
