% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_filter.R
\name{name_starts_with}
\alias{name_starts_with}
\title{scientific name starts with}
\usage{
name_starts_with(name, provider, db = td_connect, ignore_case = TRUE)
}
\description{
scientific name starts with
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }
name_contains("Homo ")
}
}
