% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{synonyms}
\alias{synonyms}
\title{synonyms}
\usage{
synonyms(name, provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb",
  "slb", "wd", "ott", "iucn"), collect = TRUE, db = td_connect())
}
\description{
Resolve provided list of names against all known synonyms
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }

sp <- c("Trochalopteron henrici gucenense",
        "Trochalopteron elliotii")
synonyms(sp)

}

}
