% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{synonyms}
\alias{synonyms}
\title{synonyms}
\usage{
synonyms(name = NULL, provider = known_providers, collect = TRUE,
  db = td_connect())
}
\arguments{
\item{name}{a character vector of scientific names, e.g. "Homo sapiens"}

\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxadb database. See details.}
}
\description{
Resolve provided list of names against all known synonyms
}
