% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{synonyms}
\alias{synonyms}
\title{synonyms}
\usage{
synonyms(name = NULL, authority = c("itis", "ncbi", "col", "tpl",
  "gbif", "fb", "slb", "wd"), collect = TRUE, db = td_connect())
}
\arguments{
\item{name}{a character vector of species names, e.g. "Homo sapiens"
(Most but not all authorities can also return ids for higher-level
taxonomic names).}

\item{authority}{from which authority should the hierachy be returned?
Default is 'itis'.}

\item{collect}{logical, default \code{TRUE}. Should we return an in-memory
data.frame (default, usually the most convenient), or a reference to
lazy-eval table on disk (useful for very large tables on which we may
first perform subsequent filtering operations.)}

\item{db}{a connection to the taxald database. See details.}
}
\description{
Resolve provided list of names against all known synonyms
}
