% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_tbl.R
\name{taxa_tbl}
\alias{taxa_tbl}
\title{Return a reference to a given table in the taxadb database}
\usage{
taxa_tbl(provider = known_providers, schema = "dwc",
  db = td_connect())
}
\arguments{
\item{provider}{from which provider should the hierachy be returned?
Default is 'itis'.}

\item{schema}{the table schema on which we want to run the query}

\item{db}{a connection to the taxadb database. Default will
attempt to connect automatically.}
}
\description{
Return a reference to a given table in the taxadb database
}
