% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_tbl.R
\name{taxa_tbl}
\alias{taxa_tbl}
\title{Return a reference to a given table in the taxald database}
\usage{
taxa_tbl(authority = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd"), schema = c("hierarchy", "taxonid", "synonyms", "common",
  "long"), db = td_connect())
}
\arguments{
\item{authority}{from which authority should the hierachy be returned?
Default is 'itis'.}

\item{schema}{the table schema on which we want to run the query}

\item{db}{a connection to the taxald database. Default will
attempt to connect automatically.}
}
\description{
Return a reference to a given table in the taxald database
}
