% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_tbl.R
\name{taxa_tbl}
\alias{taxa_tbl}
\title{Return a reference to a given table in the taxadb database}
\usage{
taxa_tbl(provider = c("itis", "ncbi", "col", "tpl", "gbif", "fb", "slb",
  "wd", "ott", "iucn"), schema = c("dwc", "common"), db = td_connect())
}
\arguments{
\item{provider}{from which provider should the hierarchy be returned?
Default is 'itis'.}

\item{schema}{the table schema on which we want to run the query}

\item{db}{a connection to the taxadb database. Default will
attempt to connect automatically.}
}
\description{
Return a reference to a given table in the taxadb database
}
\examples{
\donttest{
  \dontshow{
   ## All examples use a temporary directory
   Sys.setenv(TAXADB_HOME=tempdir())
  }

  #Clean a list of messy common names
  names <- clean_names(c("Steller's jay", "coopers Hawk"),
                binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE)

  #Get cleaned common names from a provider and search for cleaned names in that table
  taxa_tbl("itis", "common") \%>\%
  mutate_db(clean_names, "vernacularName", "vernacularNameClean",
            binomial_only = FALSE, remove_sp = FALSE, remove_punc = TRUE) \%>\%
  filter(vernacularNameClean \%in\% names)



}
}
