% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxadb_dir.R
\name{taxadb_dir}
\alias{taxadb_dir}
\title{Show the taxadb directory}
\usage{
taxadb_dir()
}
\description{
Show the taxadb directory
}
\details{
NOTE: after upgrading \code{duckdb}, a user may need to delete any
existing databases created with the previous version. An efficient
way to do so is \code{unlink(taxadb::taxadb_dir(), TRUE)}.
}
\examples{
## show the directory
taxadb_dir()
## Purge the local db
unlink(taxadb::taxadb_dir(), TRUE)

}
