% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_create.R
\name{td_create}
\alias{td_create}
\title{create a local taxonomic database}
\usage{
td_create(authorities = "itis", schema = c("hierarchy", "taxonid",
  "synonyms"), overwrite = FALSE, lines = 1e+06,
  dbdir = rappdirs::user_data_dir("taxadb"), db = td_connect(dbdir))
}
\arguments{
\item{authorities}{a list (character vector) of authorities to be included in the
database. By default, will install \code{itis}.  See details for a list of recognized
authorities. Use \code{authorities="all"} to install all available authorities automatically.}

\item{schema}{format of the database to import.}

\item{overwrite}{Should we overwrite existing tables? Default is \code{FALSE}.  Change
to "ask" for interactive interface, or \code{TRUE} to force overwrite (i.e. updating
a local database upon new release.)}

\item{lines}{number of lines that can be safely read in to memory at once. Leave
at default or increase for faster importing if you have plenty of spare RAM.}

\item{dbdir}{a location on your computer where the database should be installed.
Defaults to user data directory given by \url{rappdirs::user_data_dir}.}

\item{db}{connection to a database.  By default, taxadb will set up its own
fast \url{MonetDBLite::MonetDBLite} connection.}
}
\value{
path where database has been installed (invisibly)
}
\description{
create a local taxonomic database
}
\details{
Authorities recognized by taxadb are:
\itemize{
\item \code{itis}: Integrated Taxonomic Information System, \url{https://www.itis.gov/}
\item \code{ncbi}:  National Center for Biotechnology Information, \url{https://www.ncbi.nlm.nih.gov/taxonomy}
\item \code{col}: Catalogue of Life, \url{http://www.catalogueoflife.org/}
\item \code{tpl}: The Plant List, \url{http://www.theplantlist.org/}
\item \code{gbif}: Global Biodiversity Information Facility, \url{https://www.gbif.org/}
\item \code{fb} FishBase, \url{http://fishbase.org}
\item \code{slb}, SeaLifeBase, \url{http://sealifebase.org}
\item \code{wd}, Wikidata: \url{https://www.wikidata.org/}
\item \code{ott} OpenTree Taxonomy: \url{https://github.com/OpenTreeOfLife/reference-taxonomy}
}
}
\examples{
\dontrun{
  # tmp <- tempdir()
  # create_db(authorities = "itis", dbdir = tmp)

}
}
