% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_connect.R
\name{td_disconnect}
\alias{td_disconnect}
\title{Disconnect from the taxadb database.}
\usage{
td_disconnect(db = td_connect())
}
\arguments{
\item{db}{database connection}
}
\description{
Disconnect from the taxadb database.
}
\details{
This function manually closes a connection to the \code{taxadb} database.
}
\examples{
\donttest{

## Disconnect from the database:
td_disconnect()

}
}
